Separating Trap Files
- Separating all traps (e.g. TheoCrystalTrap)
- Removed assembly reference in favour for mod interop
