+++ everest.yaml
@@ -5,4 +5,4 @@
     - Name: Everest
       Version: 1.4896.0
     - Name: ExtendedVariantMode
-      Version: 0.40.0
+      Version: 0.40.1
